simplify jest test utils
