+++ packages/allure-jest/test/spec/labels.test.ts
@@ -1,31 +1,11 @@
-import fs from "fs";
-import { cwd } from "process";
-import { runCLI } from "@jest/core";
-import type { Config } from "@jest/types";
 import type { TestResult } from "allure-js-commons";
-import { match, stub } from "sinon";
+import { runJestTests } from "../utils";
 
 describe("labels", () => {
-  const results: TestResult[] = [];
+  let results: TestResult[];
 
   beforeEach(async () => {
-    const argv: Config.Argv = {
-      config: require.resolve("../jest.config"),
-      collectCoverage: false,
-      verbose: false,
-      silent: true,
-      $0: "",
-      _: ["./test/fixtures/labels.test.js"],
-    };
-    const writeFileSpy = stub(fs, "writeFileSync")
-      .withArgs(match("allure-results"))
-      .returns(undefined);
-
-    await runCLI(argv, [cwd()]);
-
-    writeFileSpy.args.forEach(([, rawResult]) => {
-      results.push(JSON.parse(rawResult as string) as TestResult);
-    });
+    results = await runJestTests(["./test/fixtures/labels.test.js"]);
   });
 
   it("do something", () => {
